refactor: update namespaces and folder name for tests
