+++ EmailSender.sln
@@ -10,7 +10,7 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "EmailSender.Backend.Domain"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "EmailSender.Backend.Database", "EmailSender.Backend\EmailSender.Backend.Database\EmailSender.Backend.Database.csproj", "{5E005153-349C-4653-B1F6-D6EBEFEFFB3A}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "EmailSender.IntegrationTests", "EmailSender.Tests\EmailSender.IntegrationTests\EmailSender.IntegrationTests.csproj", "{4AABF118-F0EF-4490-83D7-F91E7959D5A1}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "EmailSender.Tests.IntegrationTests", "EmailSender.Tests\EmailSender.Tests.IntegrationTests\EmailSender.Tests.IntegrationTests.csproj", "{4AABF118-F0EF-4490-83D7-F91E7959D5A1}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "EmailSender.Backend.Shared", "EmailSender.Backend\EmailSender.Backend.Shared\EmailSender.Backend.Shared.csproj", "{386D3DFC-AC2F-4233-BBEB-9C095D74C36E}"
 EndProject
@@ -18,7 +18,7 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "EmailSender.Backend.SmtpSer
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "EmailSender.Backend.Cqrs", "EmailSender.Backend\EmailSender.Backend.Cqrs\EmailSender.Backend.Cqrs.csproj", "{1165BB1A-A8D1-481B-B0E4-CDC202D66A70}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "EmailSender.UnitTests", "EmailSender.Tests\EmailSender.UnitTests\EmailSender.UnitTests.csproj", "{3A61A5DC-4906-44C5-BDEC-42D79A1CF0C6}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "EmailSender.Tests.UnitTests", "EmailSender.Tests\EmailSender.Tests.UnitTests\EmailSender.Tests.UnitTests.csproj", "{3A61A5DC-4906-44C5-BDEC-42D79A1CF0C6}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "EmailSender.Backend.UserService", "EmailSender.Services\EmailSender.Services.UserService\EmailSender.Services.UserService.csproj", "{4B8C12A8-A5A3-4E39-AEF7-D4931E89FF65}"
 EndProject
